requirements.txt
fastapi 
uvicorn 
python-multipart 
azure-cosmos 
azure-storage-blob 
azure-servicebus